requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
discord.py==2.5.2
exceptiongroup==1.2.2
fastapi==0.115.12
frozenlist==1.5.0
h11==0.14.0
idna==3.10
multidict==6.2.0
openai>=1.0.0
propcache==0.3.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyNaCl==1.5.0
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
